Fix double Tower Button Click
Once the tower button is clicked then you cannot click it again. this works however if you click it two times you can never place a tower again because the bool towerplaced will always be in a false loop. try and fix this problem. After this small problem is fixed then you can record video and move on to next step.