Fix the text for the upgrades.
The unfixed code is in the comments under placement script. The problem is that when the game runs it gives an error saying that there is no placeholder because when you click off the tower the update every  frame script in the quotation cant run which ends the game. try to fix this then test the game to see if the attack speed works.